make simulated animations part of the animation engine
Wiring up and creating required classes for simulated
animations to be part of the animation engine.

Creating a new animation node AnimationSimulationNode
for simulated animations and SimulatedAnimation class
that inherits from NumberAnimation in the animation
factory.

Also, this commit makes it possible to have two box2d
animation effects in parallel with -with previous-
option

Change-Id: I1f125df5324673e9937b8164c0fc267c9683afa0